requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.0.6
django-recaptcha==3.0.0
idna==3.4
python-environ==0.4.54
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
